requirements.txt
hydra-core==1.3.2
lightning==2.0.2
numpy==1.24.3
tabulate==0.9.0
tensorboard==2.13.0
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
wandb==0.15.2
